requirements.txt
asgiref==3.4.1
astroid==2.8.5
autopep8==1.6.0
Django==3.2.9
django-admin-rangefilter==0.8.3
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.2
pycodestyle==2.8.0
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.3
reportlab==3.6.3
sqlparse==0.4.2
toml==0.10.2
wrapt==1.13.3
